requirements.txt
fastapi
uvicorn
aiohttp
aiobotocore
boto3
aiofiles
python-dotenv